Implement Ims's VideoCallProvider, including a wrapper.
+ Implements ConnectionService.VideoCallProvider as
ImsVideoCallProviderWrapper. ConnectionService treats this as
the concrete implementation of VideoCallProvider. In detail though,
it is a wrapper around ImsVideoCallProvider which should be extended
in ImsService and implements the actual video call provider
functionality. This wrapper intermediates communication for commands
sent through Telephony and carried out in ImsService. It also
implements the callback class whose binder is provided to the
ImsVideoCallProvider.
+ ImsVideoCallProvider is an abstract class which defines
functionality which satisfies ConnectionService.VideoCallProvider.
We don't directly extend VideoCallProvider because of the
inter-process communication necessary across Telephony and
ImsService. This class provides a binder which is passed to the
ImsVideoCallProviderWrapper.
+ Implement getVideoCallProvider on ImsCallSession.

Bug: 16886403
Change-Id: I9b995baa74149b3c046789f797ab0ff49ae7165b