Fix unncessary IMS re-registration with network
When entire IMS stack turn off is going to follow the IMS
re-registration request for a specific service, there is an
unncessary overhead with extra signalling during the process of
re-registration. Also this delays the IMS turn off process.
Fixed by not performing a setFeatureValue operation for a service
turn off when the configuration allows for entire IMS stack to
be turned off.

Add Exception Handling To setAdvanced4GMode API

Bug: 22717994
Bug: 21800429
Change-Id: Ic5d7c54802818de5c8e241cacce62feb02629743