Implement plumbing for merge() related callback workflows.
Added support for a transient ImsCallSession that could eventually
turn into the underlying ImsCall when it turns into a newly minted
conference call. Also handled the series of callSessionMergeXXX()
series of call to properly set up the pluming for IMS conferences.

Also cleaned up/added some logging.

Lastly, the prototype for onCallMerged() was changed because, at this
level, there is no sensible value that we could return even
if we wanted to. No one uses this newCall parameter anyways.

Bug: 18056632
Change-Id: Ifd0e23571b8bc928f2fd365df6c2d5d0f9a65b50