Cleanup to remove misused CallGroup concept.
The ImsCall code makes extensive use of a "CallGroup".  This CallGroup has
been the source of a number of bugs, since the CallGroup concept does not
match well with how we are performing merges in IMS.

1. Removed Call Group, replacing with a "peer" and "host" references in
the ImsCall.
2. Revisited the merge logic within the context of the peer/host changes.
3. Also, changed log statements to log the ImsCall instead of the session.
This aids debugging by making it more clear what the context of the various
method operations are.

Bug: 18588383

Change-Id: I78613c44bf656c8ebf9e5d6a1145784ce630deef