Fixing disconnect tone heard when merging 3rd participant into conference.
The issue was when we set the merge peer/merge host.  We were always
setting the merge peer to the BG call.  This is fine when there is no
conference, but when we already have a conference we need to take into
account whether the FG or BG is the conference.

Bug: 18652342
Change-Id: I910a1bee3d3e04477861ed68f2a77a49d0da8471