Remove an extraneous check in callSessionMergeFailed().
There is no reason why we cannot continue with merge cleanup
(just to be safe) on a call regardless of if the "i'm in a merge"
flag is set or not. If, for some reason, the call is not
part of a merge, its a NOOP. This handles the case when the
callbacks are out of order and we receive a callSessionMergedFailed()
after a callSessionTerminate() on the merge host. It doesn't really
make sense for the callbacks to come back that way but we should
support it either way.  I also cleaned up the processMergeFailed()
function so that it doesn't abort when there is no merge peer.
Now the logic cleans up the merge host anyways and cleans up the
peer only if it exists.
Note that this is only a partial fix for this bug. The other issue
(main issue) that we need to resolve is why the first conference
puts Telecom in a state where the second conference fails.

Bug: 19852295
Change-Id: I3acf8fd4fd5c199c01ca2c4ead4119f3876e38d5