Update patch set 1
Patch Set 1:

I'll see if I can reproduce those settings.  I am not on T-Mobile so until I get a T-Mobile sim it's hard for me to do much further testing.  I did notice that when I turn wifi calling on in settings and exit and come back it is still off, but that may be due to my device not being provisioned for wifi calling and/or carrier not supporting.

When you turn it on in settings and come back to settings is it still on or off again?

Patch-set: 1