Use generic mmi error string for UT exceptions
Use default error string com.android.internal.R.string.mmiError
if no error code string received, e.g. when the network responds with a
UT error the user is presented with an error message "IMS UT exception"
which is both unclear and not localized.

STEPS TO REPRODUCE:
0) PRECONDITIONS: device and SIM Card are VoLTE enabled
1) device makes a XCAP query over the UT interface
2) The network responds with an error (409 CONFLICT in the logs)
3) Observe the error message presented to the user

Actual Result: "IMS UT exception" is presented to the user
Expected Result: a clearer and localized error message should be
presented to the user

Bug: 32792378
Change-Id: I0c71c35e7830db7de5b1d041211de54c92285b4d