Support termination of a call while merge is in process.
Added mTerminationRequestPending flag to track when a framework or user
initiated call termination is pending.  When we get a
processCallTerminated from the RIL, we can use this to override the fact
we normally bury the termination during a merge.
Note: The code change in processCallTerminated looks a lot more involved
but really just involves wrapping a try/finally block around
the main work of the method and adding the finally to cleanup state.

Bug: 22184133
Change-Id: Ie007e8f294e36fbc0eb93492089fd5521ea4ff05