Move caching of VoLTE provisioned value from ImsConfigImpl to ImsManager
ImsConfig.getVolteProvisioned() function was supposed to cache provisioned
value and return it synchronously. This CL moves this functionality to
ImsManager (and adds support for WFC and VT provisioning).

VoLTE and WFC provisioned values are asynchronously updated when
CarrierConfig changes. They are also updated when ACTION_IMS_CONFIG_CHANGED
intent is received (this intent is sent by IMS stack).

Add isWfcProvisionedOnDevice() and isVtProvisionedOnDevice() functions
to ImsManager.

Bug: 26496388
Bug: 27814311
Change-Id: I4640e5c79e339121a2899ef3cd022d2738ab244b