Ensure provisioned VT value is set appropriately on boot.
ImsManager#updateImsServiceConfig is responsible for setting the
provisioned / device VT values on boot.  Previous changes were made
to ensure that changing the enhanced 4G LTE switch would NOT change the
VT setting at the same time.  This change is made in the same spirit,
and decouples the VT feature setting on bootup from the voLTE setting.

Bug: 30454597
Change-Id: I625cfb03b7da3fc68ae52b39007ffb86613297c9