Add prototype of terminate() which supports override terminate reason code.
When ImsCall#termiante(int) is called with a reason code, the modem will
only actually disconnect the call for some disconnect reasons.

Adding an override of that method which allows the caller to specify an
override which is reported as the actual disconnect reason in the
onCallSessionTerminated callback in ImsPhoneCallTracker.

Bug: 30702393
Change-Id: I006046c03630e9d010424591f3bbb7d46da3a76f