Modify ImsVideoCallProvider to include calling UID for permission check.
Modify the ImsVideoCallProvider, which the vendor IMS stack implements to
include the UID of telephony.  The vendor IMS stack will use this when
opening the camera for the purpose of camera permission verification by
the camera service.

Test: Unit, manual.
Bug: 32747443
Merged-In: I0884cc08ea4afa23c08e3fe1e2873b3f5aac2b40
Change-Id: I0884cc08ea4afa23c08e3fe1e2873b3f5aac2b40