Modifies ImsManager to support ImsResolver
This change modifies the ImsManager to support both the old and
new versions of ImsService. This is done by creating the
ImsServiceProxyCompat class, which implements IMMTelFeature
and uses the old IImsService interface. ImsServiceProxy then
extends ImsServiceProxyCompat, which implements the new interface
for newer devices, which use IImsServiceController.

A callback interface has also been introduced to the ImsManager, which
takes information back from the ImsService regarding its status
(NOT_READY,INITIALIZING,READY) and uses it to ensure that we are not
operating on an ImsService that is not READY.

Test: Manual
Merged-In: I46f1f7237dffcdd12b66c16cd319818d1d21c101
Change-Id: I46f1f7237dffcdd12b66c16cd319818d1d21c101