Support MSIM devices in ImsManager
All configuration options dealing with IMS are
accessed via ImsManager using static methods that
always apply the settings to the user's default voice
SIM. This does not work for LTE+LTE devices that will
need to store these options on a per subscription ID
basis. This change deprecates the old methods and
introduces a "ForSlot" variant of all of the methods.

Test: Unit Tests
Change-Id: Icd18db169a4bbfdb998dd2b14d62287755f07436