Change how unpause of video is detected.
ImsVideoCallProviderWrapper was modified recently to check for the case
where the video is paused, and the video state changes to an unpaused state.
It used this to trigger a clear of the video pause tracker.
It did this by checking if the video pause tracker thought the video was
paused.  In reality, it makes more sense for a change in the videostate
from paused to unpaused to be detected, and THAT used as a basis to
clear the video pause tracker.

The previous functionality caused a problem when the user, in quick
succession:
1. Turned off the camera
2. Went to the background.

The two requests would hit the framework, causing the video pause
tracker to record the current video state as paused (as requested by
the user).  Shortly thereafter, the modem reports the fact that the
camera is off with a state of Audio RX.  This would be misinterpreted
and cause the pause tracker to be cleared.

Test: Manually tested the bug, as well as regression tested b/62784036.
Change-Id: Iaa736e93e05ef3ae5dec21cd8ebc18be464f18ae
Fixes: 63410964