Remove framework performance optimization for null video requests.
In N-MR2 we made a performance optimization to not bother sending
no-op video requests to the vendor IMS layer.  In concept this performance
optimization makes sense; if a higher layer tells us to change from
the current video state to the current video state, then don't send it.

It turns out that the vendor IMS layer does not interpret the delta
style video requests as they were intended.

To fix this, we'll remove the performance optization and opt to make the
contract between the framework and the vendor IMS layer more explicit in
the future to avoid this kind of problem.

This fix corrects an issue where:
A and B are in a video call.
A pauses the video
B pauses the video
A un-pauses the video

When A un-pauses the video B would also unpause, even though it should
not.  As it turns out the performance optimization was resulting in B
not issuing the pause request to the modem.

Test: Manual test
Change-Id: Ia29471b3f13d5b104affcf634e1315fb7c71451b
Fixes: 62954065