Filter resume requests when video is already resumed.
If the request being sent to the video provider is to resume the video,
but the video stream is already resumed, do not send this request to the
modem.

In the case of swapping between calls, the video appears to resume
automatically on the modem-side, so sending another resume request is
redundant.

Also, in VideoPauseTracker, correcting the case where we get a resume
request, but there are no remaining pause requests in the tracker.
Although that case shouldn't run up in reality, if it did we should still
let the resume request pass along (since it would otherwise be re-written
as a pause).

Test: Manual test.
Change-Id: Ib9b9acaf2d92b1485e4766a13701fd472d6c117d
Fixes: 63606238