Prevent IMS call merge if the peer call has a pending update request.
Currently, if the user requests merge, and the call they request the merge
on is in the midst of some other pending operation (e.g. hold, resume), the
system will prevent the merge from taking place.  The logic, however does
not account for the fact that the background call being merged into the
foreground call could also be engaged in some other update operation.

Added code to ensure that both calls are not actively engaged in some
pending operation.

Test: Manual
Change-Id: Icc2f3786865345435bbf138b54736fc6f09c0aa4
Fixes: 63764631