Integrate Registration APIs
Integrates IMS Registration APIs into the platform
and fixes a few issues that were occuring when
switching between two different ImsServices:

1) Preserve Backwards compatibility by modifying
ImsConnectionStateListener to extend the new
ImsRegistration Listener APIs.

2) Add a new callback (ImsFeatureUpdate#notifyUnavailable)
that is triggered by the framework when the ImsService
is no longer available due to it switching or crashing.

3) Small refactor of ImsServiceProxy for initialization and
querying ImsRegistration.

4) Integrate ImsRegistration into ImsManager.

Bug: 63987047
Test: Manual, Telephony unit tests
Change-Id: I268efbfc3df7e400a3348f9b4586edda53ca728a