Fix editable roaming mode query bug
We were querying whether or not the WFC mode was editable,
instead of the WFC roaming mode. This was causing the issue
where the user was unable to set WiFi preferred mode when
the WFC mode was not editable.

Bug: 80168790
Test: Manual
Change-Id: I8895ee97e7917192305e8ad53fd4de9ce1bfca01
Merged-In: I5f219cdab2bea5ac344252fc0fc5ee849a9fa937