Fix SubscriptionController crash during subid verification
Receiving reports of ImsManager causing crashes when trying to
update settings in the siminfo db related to WFC. This should
only be triggered when there is a valid sub id, which causes
a crash. There is not enough info in the bugreports to find
the root cause, but I can at least fix the phone crash and log an
error when this happens.

Test: atest FrameworksTelephonyTests
Bug: 113291931
Change-Id: I825392c6a8ac1602ac7e5c915b80b41aa6734d27