Speculative fix for ImsManager crashes.
It turns out the speculative fix in b/113291931 wasn't capturing the
cases which are causing crashes.

Dug into the code and realized that the checks performed in
SubscriptionController go further than what was done here previously.
Replicating the same checks her and logging actual sub Id so we can
perhaps get more information as to why this is failing in the field.

Test: Build
Bug: 115894442
Change-Id: Ifd4fe110ae797c435ce737845af99218fdab5401